+++ packages/orchestrator-ui-components/src/components/WfoForms/formFields/SplitPrefixStyling.ts
@@ -1,4 +1,4 @@
-import { css } from '@emotion/core';
+import { css } from '@emotion/react';
 
 export const splitPrefixStyling = css`
     section.split-prefix {
