+++ packages/orchestrator-ui-components/src/components/WfoForms/formFields/surf/types.ts
@@ -27,3 +27,27 @@ export interface ImsNode {
     name: string;
     status: string;
 }
+
+export interface SortOption<nameStrings = string> {
+    name: nameStrings;
+    descending: boolean;
+}
+
+export interface IpPrefix {
+    id: number;
+    prefix: string;
+    version: number;
+}
+
+export interface IpBlock {
+    id: number;
+    prefix: string;
+    ip_network: string;
+    description: string;
+    state: number;
+    parent: number;
+    version: number;
+    parent_ipam_id: number;
+    is_subnet: boolean;
+    state_repr: string;
+}
