+++ packages/orchestrator-ui-components/src/components/WfoForms/formFields/surf/utils.ts
@@ -1,0 +1,1 @@
+export const ipamStates = ['Free', 'Allocated', null, 'Planned', null, null];
