+++ packages/orchestrator-ui-components/src/components/WfoForms/formFields/types.ts
@@ -40,38 +40,6 @@ export interface Option<Value = string> {
     label: string;
 }
 
-export interface ServicePort {
-    subscription_id?: string;
-    vlan?: string;
-    bandwidth?: number;
-    nonremovable?: boolean;
-    modifiable?: boolean;
-}
-
-export interface IpPrefix {
-    id: number;
-    prefix: string;
-    version: number;
-}
-
-export interface IpBlock {
-    id: number;
-    prefix: string;
-    ip_network: string;
-    description: string;
-    state: number;
-    parent: number;
-    version: number;
-    parent_ipam_id: number;
-    is_subnet: boolean;
-    state_repr: string;
-}
-
-export interface SortOption<nameStrings = string> {
-    name: nameStrings;
-    descending: boolean;
-}
-
 export interface Option<Value = string> {
     value: Value;
     label: string;
